Refactor: 사이드바 컴포넌트들 폴더 구조 변경
